requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
python-dotenv
werkzeug
openai
Pillow
gunicorn
psycopg2-binary